==> 3_merge-flats-and-house.ipynb <==
import numpy as np
import pandas as pd

flats = pd.read_csv('agarwal_flats_cleaned.csv')
houses = pd.read_csv('agarwal_house_cleaned.csv')

df = pd.concat([flats,houses],ignore_index=True)

df = df.sample(df.shape[0],ignore_index=True)

df.head()

df.to_csv('agarwal_gurgaon_properties.csv',index=False)
