==> 7_eda-pandas-profiling.ipynb <==
import pandas as pd
from ydata_profiling import ProfileReport
# pandas profiling does performs Automated EDA

# Load your dataset
df = pd.read_csv('agarwal_gurgaon_properties_cleaned_v2.csv').drop_duplicates()

# Create the ProfileReport object
profile = ProfileReport(df, title='Pandas Profiling Report', explorative=True)

# Generate the report
profile.to_file("agarwal_output_report.html")


!pip install pandas-profiling


pip install -U ydata-profiling
